WAITP-1142 fix issues with image upload

+++ packages/utils/src/s3/S3Client.js
@@ -100,6 +100,7 @@ export class S3Client {
 
   async deleteFile(filePath) {
     const fileName = filePath.split(getS3ImageFilePath())[1];
+    if (!(await this.checkIfFileExists(fileName))) return null;
     return new Promise((resolve, reject) => {
       this.s3.deleteObject(
         {
